mplayer: fix buffer handling during channel change.
Without this change, when changing from, for example, stereo
to 5.1, there may be a little stereo data left in the output
buffer when the audio output is reinitialized with the new
format. If that little stereo data is not a multiple of 6 samples
it causes a shift in the mapping of the 5.1 channels.

With this patch, decoded audio data in new format is not filtered
as long as there are still samples in old format that have not
yet been sent to the AO driver.

Based on a suggestion by Reimar Döffinger.