vf_bmovl: Try to init and access overlay buffers correctly
The overlay holds the image that is blended onto each video frame.
It can potentially be updated on each frame to be processed.

The overlay is stored as separate buffers for luma, alpha and chroma
planes. The luma and alpha planes have same resolution as the frame
of the playing video and chroma planes have halved resolution.

Only access this buffers with that dimensions, do not use the
current video frame's stride to access these internal buffers.
They always have a size of width * height or width * height / 4 .

Partially based on patch by floemker (see Trac ticket #2304)

Fixes #2304